Refactor `bytes` module.
Replaces the `Serializer` and `Deserializer` types with `Bytes` and
`BytesMut` from the [bytes] crate, and adapts the serialization traits
accordingly.

This keeps us from having to re-invent the wheel and also simplifies the
usage of this module.

Also implements serialization of the SOME/IP `Message` type, as promised
in `051f8f1`.

[bytes]: https://docs.rs/bytes/latest/bytes/