Add `endpoint` module. (#7)
* Simplify `socket` module.

These changes remove a lot of unnecessary code, which were making the
architecture of the `socket` module needlessly complex.

* Remove `net` module.

Useful types were moved to the `support` module.

* Improve `ipv4!` macro.

Adds an option without any arguments that generates a random address in
the range `127.0.0.1-127.0.0.255` with a random port in the range
`49_152..65_535`.

* Simplify `someip` module.

Removes a lot of unnecessary code.

Serialization will be handled in a future commit.

* Refactor `bytes` module.

Replaces the `Serializer` and `Deserializer` types with `Bytes` and
`BytesMut` from the [bytes] crate, and adapts the serialization traits
accordingly.

This keeps us from having to re-invent the wheel and also simplifies the
usage of this module.

Also implements serialization of the SOME/IP `Message` type, as promised
in `051f8f1`.

[bytes]: https://docs.rs/bytes/latest/bytes/

* Add `endpoint` module.

This module defines an interface for implementing SOME/IP endpoints, and
provides a concrete implementation for Version 1 of the protocol.

Also, adds a placeholder `error` module to lay the ground work for
centralized error handling.